Fix debian keyserver address and place debian.cnf
The debian keyserver address was not complete and causing errors when adding
the package repo.

Also added the template for /etc/mysql/debian.cnf in. Not sure if it got
removed in the merge resolution process or something else, but without it
ubuntu generates a random password which is tries to use to connect for the
debian_sys_maint account, which fails since we set it from attributes.